flask_cloudfirestore/requirements.txt
alembic==1.12.0
blinker==1.6.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.5
colorama==0.4.6
cryptography==41.0.4
dnspython==2.5.0
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-MySQL==1.5.2
Flask-SQLAlchemy==3.1.1
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
googleapis-common-protos==1.62.0
# greenlet==2.0.2
greenlet
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.1
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.1
PyMySQL==1.1.0
requests==2.31.0
rsa==4.9
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==2.3.7
zipp==3.17.0
